# docs/foccacia-api-spec.yaml
openapi: 3.0.0
info:
  title: FOCCACIA API
  version: 1.0.0
  description: API for managing football clubs, leagues, and user groups.
servers:
  - url: http://localhost:3000/api
    description: Local development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
paths:
  /teams/search:
    get:
      summary: Search clubs by name
      parameters:
        - name: name
          in: query
          required: true
          description: Name of the club to search for.
          schema:
            type: string
      responses:
        "200":
          description: Successful response with team data.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    team:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 33
                        name:
                          type: string
                          example: Real Madrid
                        logo:
                          type: string
                          example: https://media.api-sports.io/football/teams/33.png
                    venue:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: Santiago Bernabeu
                        city:
                          type: string
                          example: Madrid
                        capacity:
                          type: integer
                          example: 81044
        "400":
          description: Missing or invalid query parameter.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Team name is required.
        "404":
          description: No teams found with the given name.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No teams found with the given name.
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to fetch team data.
...
  /leagues/search:
    get:
      summary: Search leagues by team
      parameters:
        - name: team
          in: query
          required: true
          description: ID of the team to search leagues for.
          schema:
            type: integer
      responses:
        "200":
          description: List of leagues for the team.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      example: 140
                    name:
                      type: string
                      example: La Liga
                    season:
                      type: integer
                      example: 2023
        "404":
          description: No leagues found for the given team ID.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: No leagues found.
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Internal server error.
  /groups:
    post:
      summary: Create a new group
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: My Favorite Teams
                description:
                  type: string
                  example: A collection of my favorite teams.
      responses:
        "401":
          description: Unauthorized (missing or invalid token).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized.
        "201":
          description: Group created successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: group-12345
                  name:
                    type: string
                    example: My Favorite Teams
                  description:
                    type: string
                    example: A collection of my favorite teams.
                  teams:
                    type: array
                    items:
                      type: object
                      properties:
                        teamName:
                          type: string
                          example: Real Madrid
                        stadiumName:
                          type: string
                          example: Santiago Bernabeu
                        leagueName:
                          type: string
                          example: La Liga
                        season:
                          type: integer
                          example: 2023
    get:
      summary: List all groups
      security:
        - bearerAuth: []
      responses:
        "401":
          description: Unauthorized (missing or invalid token).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Unauthorized.
        "200":
          description: List of all groups.
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: group-12345
                    name:
                      type: string
                      example: My Favorite Teams
                    description:
                      type: string
                      example: A collection of my favorite teams.
...